Move pick tests to the pick module.
