+++ src/pugbot/models/game.rs
@@ -393,85 +393,4 @@ mod tests {
 
     assert_eq!(game.phase, Some(Phases::PlayerDrafting));
   }
-
-  #[test]
-  fn test_pick_player() {
-    let authors: Vec<User> = p!(Vec, "authors");
-    let (team_count, team_size) = (2, (authors.len() / 2) as u32);
-    // Choosing 2 teams of 5 here since there are 10 authors in authors.json
-    let game = &mut Game::new(
-      vec![],
-      DraftPool::new(authors, team_count * team_size),
-      1,
-      Vec::new(),
-      team_count,
-      team_size,
-    );
-    game.next_phase();
-    // Check that captain-selection has finished before proceeding as this is a
-    // precondition of the test.
-    assert_eq!(game.select_captains(), Ok(()));
-
-    // Make a random selection from available players
-    let pool = game.draft_pool.available_players.clone();
-
-    if let Some(key) = pool.keys().next() {
-      if let Some(_user) = game.draft_pool.available_players.get(key) {
-        let message = p!(Message, "message");
-        // Drafting a single player works as expected?
-        assert_eq!(
-          commands::pick::draft_player(game, &message, false, *key),
-          Ok(())
-        );
-      }
-    }
-
-    // There should be as many teams as specified.
-    assert_eq!(game.teams.len() as u32, team_count);
-    // The order in which teams choose seems to be non-deterministic, so instead
-    // of checking that team 1 has x members and team 2 has x+1 members, just
-    // test their member counts equal 3.
-    let member_count = &game
-      .teams
-      .iter()
-      .fold(0, |acc, team| acc + team.members.len());
-    assert_eq!(*member_count as u32, 3);
-  }
-
-  #[test]
-  fn test_full_teams() {
-    let authors: Vec<User> = p!(Vec, "authors");
-    let message = p!(Message, "message");
-    let (team_count, team_size) = (2, (authors.len() / 2) as u32);
-    let game = &mut Game::new(
-      vec![],
-      DraftPool::new(authors, team_count * team_size),
-      1,
-      Vec::new(),
-      team_count,
-      team_size,
-    );
-    game.next_phase();
-    assert_eq!(game.select_captains(), Ok(()));
-    let player_pool = game.draft_pool.available_players.clone();
-    for (key, _) in player_pool.iter() {
-      assert_eq!(
-        commands::pick::draft_player(game, &message, false, *key),
-        Ok(())
-      );
-    }
-    // available_players should be empty. Each drafted player is popped out of
-    // the available_players pool.
-    assert_eq!(game.draft_pool.available_players.len(), 0);
-    // Since all players were drafted and teams are now full, the game should
-    // proceed to the next phase.
-    assert_eq!(game.phase, Some(Phases::MapSelection));
-    let member_count = &game
-      .teams
-      .iter()
-      .fold(0, |acc, team| acc + team.members.len());
-    // Final post-condition: The sum of the counts of each team's roster should
-    // equal the max size of the draft pool.
-    assert_eq!(*member_count as u32, team_count * team_size);
-  }
 }
